Refactor api-lib.php into modular libraries and centralize validation logic.
- Created `lib/common.php` for shared utilities.
- Updated `lib/validation.php` with centralized input validation and sanitization functions.
- Implemented `validate_appointment_payload` for consistent appointment validation.
- Modularized `api-lib.php` logic into `lib/models.php`, `lib/business.php`, `lib/storage.php`, `lib/audit.php`, `lib/auth.php`, `lib/http.php`, `lib/email.php`.
- Updated `api-lib.php` to serve as a bootstrapper requiring the new modules.
- Refactored `api.php` to use the centralized validation logic.
