+++ lib/audit.php
@@ -1,6 +1,9 @@
 <?php
 declare(strict_types=1);
 
+require_once __DIR__ . '/common.php';
+require_once __DIR__ . '/storage.php';
+
 /**
  * Audit Logging Helpers
  */
