+++ lib/email.php
@@ -1,6 +1,9 @@
 <?php
 declare(strict_types=1);
 
+require_once __DIR__ . '/common.php';
+require_once __DIR__ . '/models.php';
+
 /**
  * Email sending logic.
  */
@@ -36,143 +39,46 @@ function smtp_send_mail(string $to, string $subject, string $body): bool
         return false;
     }
 
-    $from = $cfg['from'] !== '' ? $cfg['from'] : $cfg['user'];
-    $fromName = $cfg['from_name'];
-
-    $socket = @fsockopen($cfg['host'], $cfg['port'], $errno, $errstr, 10);
-    if (!$socket) {
-        error_log("Piel en Armonía SMTP: no se pudo conectar a {$cfg['host']}:{$cfg['port']} - {$errstr}");
+    // Verificar si PHPMailer está disponible
+    if (!class_exists('PHPMailer\PHPMailer\PHPMailer')) {
+        error_log('Piel en Armonía: PHPMailer no encontrado. Ejecuta "composer install".');
         return false;
     }
-    stream_set_timeout($socket, 15);
 
-    $log = [];
-
-    $readLine = static function () use ($socket, &$log): string {
-        $response = '';
-        while (($line = fgets($socket, 512)) !== false) {
-            $response .= $line;
-            if (isset($line[3]) && $line[3] === ' ') {
-                break;
-            }
-        }
-        $log[] = 'S: ' . trim($response);
-        return $response;
-    };
-
-    $sendCmd = static function (string $cmd) use ($socket, $readLine, &$log): string {
-        $log[] = 'C: ' . trim($cmd);
-        fwrite($socket, $cmd . "\r\n");
-        return $readLine();
-    };
+    $mail = new PHPMailer\PHPMailer\PHPMailer(true);
 
     try {
-        // Greeting
-        $greeting = $readLine();
-        if (strpos($greeting, '220') !== 0) {
-            error_log('Piel en Armonía SMTP: saludo inesperado: ' . trim($greeting));
-            fclose($socket);
-            return false;
-        }
+        // Configuración del servidor
+        $mail->isSMTP();
+        $mail->Host       = $cfg['host'];
+        $mail->SMTPAuth   = true;
+        $mail->Username   = $cfg['user'];
+        $mail->Password   = $cfg['pass'];
+        // STARTTLS explícito como en la implementación original (puerto 587)
+        $mail->SMTPSecure = PHPMailer\PHPMailer\PHPMailer::ENCRYPTION_STARTTLS;
+        $mail->Port       = $cfg['port'];
+        $mail->CharSet    = 'UTF-8';
 
-        // EHLO
-        $sendCmd('EHLO pielarmonia.com');
+        // Remitente y destinatario
+        $from = $cfg['from'] !== '' ? $cfg['from'] : $cfg['user'];
+        $fromName = $cfg['from_name'];
 
-        // STARTTLS
-        $tlsResp = $sendCmd('STARTTLS');
-        if (strpos($tlsResp, '220') !== 0) {
-            error_log('Piel en Armonía SMTP: STARTTLS falló: ' . trim($tlsResp));
-            fclose($socket);
-            return false;
-        }
+        $mail->setFrom($from, $fromName);
+        $mail->addAddress($to);
 
-        $cryptoOk = stream_socket_enable_crypto($socket, true, STREAM_CRYPTO_METHOD_TLSv1_2_CLIENT | STREAM_CRYPTO_METHOD_TLSv1_3_CLIENT);
-        if (!$cryptoOk) {
-            error_log('Piel en Armonía SMTP: fallo al habilitar TLS');
-            fclose($socket);
-            return false;
-        }
+        // Contenido
+        $mail->isHTML(false);
+        $mail->Subject = $subject;
+        $mail->Body    = $body;
 
-        // Re-EHLO after TLS
-        $sendCmd('EHLO pielarmonia.com');
-
-        // AUTH LOGIN
-        $authResp = $sendCmd('AUTH LOGIN');
-        if (strpos($authResp, '334') !== 0) {
-            error_log('Piel en Armonía SMTP: AUTH LOGIN no aceptado: ' . trim($authResp));
-            fclose($socket);
-            return false;
-        }
-
-        $userResp = $sendCmd(base64_encode($cfg['user']));
-        if (strpos($userResp, '334') !== 0) {
-            error_log('Piel en Armonía SMTP: usuario rechazado');
-            fclose($socket);
-            return false;
-        }
-
-        $passResp = $sendCmd(base64_encode($cfg['pass']));
-        if (strpos($passResp, '235') !== 0) {
-            error_log('Piel en Armonía SMTP: autenticación falló - verifica contraseña de aplicación');
-            fclose($socket);
-            return false;
-        }
-
-        // MAIL FROM
-        $fromResp = $sendCmd("MAIL FROM:<{$from}>");
-        if (strpos($fromResp, '250') !== 0) {
-            error_log('Piel en Armonía SMTP: MAIL FROM rechazado: ' . trim($fromResp));
-            fclose($socket);
-            return false;
-        }
-
-        // RCPT TO
-        $rcptResp = $sendCmd("RCPT TO:<{$to}>");
-        if (strpos($rcptResp, '250') !== 0) {
-            error_log('Piel en Armonía SMTP: RCPT TO rechazado: ' . trim($rcptResp));
-            fclose($socket);
-            return false;
-        }
-
-        // DATA
-        $dataResp = $sendCmd('DATA');
-        if (strpos($dataResp, '354') !== 0) {
-            error_log('Piel en Armonía SMTP: DATA rechazado: ' . trim($dataResp));
-            fclose($socket);
-            return false;
-        }
-
-        // Construir mensaje
-        $encodedSubject = '=?UTF-8?B?' . base64_encode($subject) . '?=';
-        $encodedFrom = '=?UTF-8?B?' . base64_encode($fromName) . '?= <' . $from . '>';
-        $messageId = '<' . bin2hex(random_bytes(16)) . '@pielarmonia.com>';
-
-        $headers = "From: {$encodedFrom}\r\n";
-        $headers .= "To: {$to}\r\n";
-        $headers .= "Subject: {$encodedSubject}\r\n";
-        $headers .= "MIME-Version: 1.0\r\n";
-        $headers .= "Content-Type: text/plain; charset=UTF-8\r\n";
-        $headers .= "Content-Transfer-Encoding: base64\r\n";
-        $headers .= "Message-ID: {$messageId}\r\n";
-        $headers .= "Date: " . date('r') . "\r\n";
-
-        $encodedBody = chunk_split(base64_encode($body), 76, "\r\n");
-        $fullMessage = $headers . "\r\n" . $encodedBody . "\r\n.";
-
-        $endResp = $sendCmd($fullMessage);
-        if (strpos($endResp, '250') !== 0) {
-            error_log('Piel en Armonía SMTP: mensaje no aceptado: ' . trim($endResp));
-            fclose($socket);
-            return false;
-        }
-
-        $sendCmd('QUIT');
-        fclose($socket);
+        $mail->send();
         return true;
 
+    } catch (PHPMailer\PHPMailer\Exception $e) {
+        error_log("Piel en Armonía SMTP: Error al enviar - {$mail->ErrorInfo}");
+        return false;
     } catch (Throwable $e) {
-        error_log('Piel en Armonía SMTP: excepción - ' . $e->getMessage());
-        @fclose($socket);
+        error_log("Piel en Armonía SMTP: Excepción general - {$e->getMessage()}");
         return false;
     }
 }
@@ -207,75 +113,31 @@ function maybe_send_appointment_email(array $appointment): bool
         return false;
     }
 
-    $clinic = 'Piel en Armonia';
-    $name = $appointment['name'] ?? 'paciente';
-    $service = get_service_label((string) ($appointment['service'] ?? ''));
-    $doctor = get_doctor_label((string) ($appointment['doctor'] ?? ''));
-    $date = format_date_label((string) ($appointment['date'] ?? ''));
-    $rawDate = (string) ($appointment['date'] ?? '');
-    $time = (string) ($appointment['time'] ?? '-');
-    $price = (string) ($appointment['price'] ?? get_service_total_price((string) ($appointment['service'] ?? '')));
-    $payment = get_payment_method_label((string) ($appointment['paymentMethod'] ?? 'unpaid'));
-    $paymentStatus = get_payment_status_label((string) ($appointment['paymentStatus'] ?? 'pending'));
+    $clinicName = 'Piel en Armonía';
+    $subject = 'Confirmacion de cita - ' . $clinicName;
+    $message = "Hola " . ($appointment['name'] ?? 'paciente') . ",\n\n";
+    $message .= "Tu cita fue registrada correctamente.\n";
 
-    $subject = 'Confirmacion de cita | ' . $rawDate . ' ' . $time . ' | ' . $clinic;
+    $serviceLabel = function_exists('get_service_label') ? get_service_label((string)($appointment['service'] ?? '')) : ($appointment['service'] ?? '-');
+    $doctorLabel = function_exists('get_doctor_label') ? get_doctor_label((string)($appointment['doctor'] ?? '')) : ($appointment['doctor'] ?? '-');
+    $dateLabel = function_exists('format_date_label') ? format_date_label((string)($appointment['date'] ?? '')) : ($appointment['date'] ?? '-');
+    $paymentStatusLabel = function_exists('get_payment_status_label') ? get_payment_status_label((string)($appointment['paymentStatus'] ?? 'pending')) : ($appointment['paymentStatus'] ?? 'pending');
 
-    $m = "";
-    $m .= "PIEL EN ARMONIA\n";
-    $m .= "Dermatologia Especializada\n";
-    $m .= str_repeat("-", 45) . "\n\n";
-
-    $m .= "Hola " . $name . ",\n\n";
-    $m .= "Tu cita ha sido registrada exitosamente.\n";
-    $m .= "A continuacion los detalles:\n\n";
-
-    $m .= str_repeat("-", 45) . "\n";
-    $m .= "  DETALLES DE TU CITA\n";
-    $m .= str_repeat("-", 45) . "\n\n";
-    $m .= "  Servicio:     " . $service . "\n";
-    $m .= "  Doctor:       " . $doctor . "\n";
-    $m .= "  Fecha:        " . $date . "\n";
-    $m .= "  Hora:         " . $time . "\n";
-    $m .= "  Precio:       " . $price . "\n";
-    $m .= "  Pago:         " . $payment . "\n";
-    $m .= "  Estado:       " . $paymentStatus . "\n\n";
-
-    $m .= str_repeat("-", 45) . "\n";
-    $m .= "  ANTES DE TU CITA\n";
-    $m .= str_repeat("-", 45) . "\n\n";
-    $m .= "  * Llega 10 minutos antes de tu hora.\n";
-    $m .= "  * Trae tu cedula o documento de identidad.\n";
-    $m .= "  * Si tienes examenes previos, trailos a la consulta.\n\n";
-
-    $m .= str_repeat("-", 45) . "\n";
-    $m .= "  COMO LLEGAR\n";
-    $m .= str_repeat("-", 45) . "\n\n";
-    $m .= "  Direccion:\n";
-    $m .= "  " . clinic_address_display() . "\n\n";
-    $m .= "  Referencia: Frente al Colegio de las Mercedarias,\n";
-    $m .= "  a 2 cuadras de la Maternidad Isidro Ayora.\n\n";
-    $m .= "  Ver en Google Maps:\n";
-    $m .= "  " . clinic_map_url() . "\n\n";
-
-    $m .= "  Horario de atencion:\n";
-    $m .= "  " . clinic_schedule_display() . "\n\n";
+    $message .= "Servicio: " . $serviceLabel . "\n";
+    $message .= "Doctor: " . $doctorLabel . "\n";
+    $message .= "Fecha: " . $dateLabel . " (" . ($appointment['date'] ?? '') . ")\n";
+    $message .= "Hora: " . ($appointment['time'] ?? '-') . "\n";
+    $message .= "Estado de pago: " . $paymentStatusLabel . "\n\n";
 
     $token = $appointment['rescheduleToken'] ?? '';
     if ($token !== '') {
-        $m .= str_repeat("-", 45) . "\n";
-        $m .= "  REPROGRAMAR O CANCELAR\n";
-        $m .= str_repeat("-", 45) . "\n\n";
-        $m .= "  Si necesitas cambiar la fecha u hora:\n";
-        $m .= "  https://pielarmonia.com/?reschedule=" . $token . "\n\n";
+        $message .= "Si necesitas reprogramar tu cita, usa este enlace:\n";
+        $message .= "https://pielarmonia.com/?reschedule=" . $token . "\n\n";
     }
 
-    $m .= str_repeat("-", 45) . "\n\n";
-    $m .= "Gracias por confiar en Piel en Armonia.\n";
-    $m .= "Estamos para cuidarte.\n\n";
-    $m .= "Telefono/WhatsApp: " . clinic_whatsapp_display() . "\n";
-    $m .= "Web: https://pielarmonia.com\n";
+    $message .= "Gracias por confiar en nosotros.";
 
-    return send_mail($to, $subject, $m);
+    return send_mail($to, $subject, $message);
 }
 
 function maybe_send_admin_notification(array $appointment): bool
@@ -290,89 +152,39 @@ function maybe_send_admin_notification(array $appointment): bool
         return false;
     }
 
-    $name = $appointment['name'] ?? '-';
-    $email = $appointment['email'] ?? '-';
-    $phone = $appointment['phone'] ?? '-';
-    $service = get_service_label((string) ($appointment['service'] ?? ''));
-    $doctor = get_doctor_label((string) ($appointment['doctor'] ?? ''));
-    $date = format_date_label((string) ($appointment['date'] ?? ''));
-    $rawDate = (string) ($appointment['date'] ?? '');
-    $time = (string) ($appointment['time'] ?? '-');
-    $price = (string) ($appointment['price'] ?? get_service_total_price((string) ($appointment['service'] ?? '')));
-    $payment = get_payment_method_label((string) ($appointment['paymentMethod'] ?? 'unpaid'));
-    $paymentStatus = get_payment_status_label((string) ($appointment['paymentStatus'] ?? 'pending'));
+    $serviceLabel = function_exists('get_service_label') ? get_service_label((string)($appointment['service'] ?? '')) : ($appointment['service'] ?? '-');
+    $doctorLabel = function_exists('get_doctor_label') ? get_doctor_label((string)($appointment['doctor'] ?? '')) : ($appointment['doctor'] ?? '-');
 
-    $subject = '[Piel en Armonia] Nueva cita: ' . $name . ' | ' . $rawDate . ' ' . $time;
-
-    $b = "";
-    $b .= "PIEL EN ARMONIA - NOTIFICACION DE CITA\n";
-    $b .= str_repeat("=", 50) . "\n\n";
-
-    $b .= "Se ha registrado una nueva cita desde la web.\n\n";
-
-    $b .= str_repeat("-", 50) . "\n";
-    $b .= "  PACIENTE\n";
-    $b .= str_repeat("-", 50) . "\n";
-    $b .= "  Nombre:      " . $name . "\n";
-    $b .= "  Email:       " . $email . "\n";
-    $b .= "  Telefono:    " . $phone . "\n\n";
-
-    $b .= str_repeat("-", 50) . "\n";
-    $b .= "  CITA\n";
-    $b .= str_repeat("-", 50) . "\n";
-    $b .= "  Servicio:    " . $service . "\n";
-    $b .= "  Doctor:      " . $doctor . "\n";
-    $b .= "  Fecha:       " . $date . "\n";
-    $b .= "  Hora:        " . $time . "\n";
-    $b .= "  Precio:      " . $price . "\n\n";
-
-    $b .= str_repeat("-", 50) . "\n";
-    $b .= "  PAGO\n";
-    $b .= str_repeat("-", 50) . "\n";
-    $b .= "  Metodo:      " . $payment . "\n";
-    $b .= "  Estado:      " . $paymentStatus . "\n\n";
-
-    $reason = trim((string) ($appointment['reason'] ?? ''));
-    $area = trim((string) ($appointment['affectedArea'] ?? ''));
-    $evolution = trim((string) ($appointment['evolutionTime'] ?? ''));
-    if ($reason !== '' || $area !== '' || $evolution !== '') {
-        $b .= str_repeat("-", 50) . "\n";
-        $b .= "  INFORMACION CLINICA\n";
-        $b .= str_repeat("-", 50) . "\n";
-        if ($reason !== '') {
-            $b .= "  Motivo:              " . $reason . "\n";
-        }
-        if ($area !== '') {
-            $b .= "  Zona afectada:       " . $area . "\n";
-        }
-        if ($evolution !== '') {
-            $b .= "  Tiempo de evolucion: " . $evolution . "\n";
-        }
-        $b .= "\n";
-    }
-
-    $photoCount = (int) ($appointment['casePhotoCount'] ?? 0);
-    if ($photoCount > 0) {
-        $b .= str_repeat("-", 50) . "\n";
-        $b .= "  FOTOS ADJUNTAS (" . $photoCount . ")\n";
-        $b .= str_repeat("-", 50) . "\n";
-        if (isset($appointment['casePhotoUrls']) && is_array($appointment['casePhotoUrls'])) {
-            foreach ($appointment['casePhotoUrls'] as $photoUrl) {
-                $url = trim((string) $photoUrl);
-                if ($url !== '') {
-                    $b .= "  " . $url . "\n";
-                }
+    $clinicName = 'Piel en Armonía';
+    $subject = 'Nueva cita agendada - ' . $clinicName;
+    $body = "Se ha agendado una nueva cita:\n\n";
+    $body .= "Paciente: " . ($appointment['name'] ?? '-') . "\n";
+    $body .= "Email: " . ($appointment['email'] ?? '-') . "\n";
+    $body .= "Telefono: " . ($appointment['phone'] ?? '-') . "\n";
+    $body .= "Motivo: " . ($appointment['reason'] ?? '-') . "\n";
+    $body .= "Zona: " . ($appointment['affectedArea'] ?? '-') . "\n";
+    $body .= "Evolucion: " . ($appointment['evolutionTime'] ?? '-') . "\n";
+    $body .= "Consentimiento datos: " . ((isset($appointment['privacyConsent']) && $appointment['privacyConsent']) ? 'si' : 'no') . "\n";
+    $body .= "Servicio: " . $serviceLabel . "\n";
+    $body .= "Doctor: " . $doctorLabel . "\n";
+    $body .= "Fecha: " . ($appointment['date'] ?? '-') . "\n";
+    $body .= "Hora: " . ($appointment['time'] ?? '-') . "\n";
+    $body .= "Precio: " . ($appointment['price'] ?? '-') . "\n";
+    $body .= "Metodo de pago: " . ($appointment['paymentMethod'] ?? '-') . "\n";
+    $body .= "Estado de pago: " . ($appointment['paymentStatus'] ?? '-') . "\n";
+    $body .= "Fotos adjuntas: " . (int) ($appointment['casePhotoCount'] ?? 0) . "\n";
+    if (isset($appointment['casePhotoUrls']) && is_array($appointment['casePhotoUrls']) && count($appointment['casePhotoUrls']) > 0) {
+        $body .= "URLs de fotos:\n";
+        foreach ($appointment['casePhotoUrls'] as $photoUrl) {
+            $url = trim((string) $photoUrl);
+            if ($url !== '') {
+                $body .= "- " . $url . "\n";
             }
         }
-        $b .= "\n";
     }
+    $body .= "\nFecha de registro: " . local_date('d/m/Y H:i') . "\n";
 
-    $b .= str_repeat("=", 50) . "\n";
-    $b .= "Registrado: " . local_date('d/m/Y H:i') . "\n";
-    $b .= "Consentimiento datos: " . ((isset($appointment['privacyConsent']) && $appointment['privacyConsent']) ? 'Si' : 'No') . "\n";
-    $b .= "Ubicacion: " . clinic_map_url() . "\n";
-
-    return send_mail($adminEmail, $subject, $b);
+    return send_mail($adminEmail, $subject, $body);
 }
 
 function maybe_send_cancellation_email(array $appointment): bool
@@ -382,16 +194,20 @@ function maybe_send_cancellation_email(array $appointment): bool
         return false;
     }
 
+    $serviceLabel = function_exists('get_service_label') ? get_service_label((string)($appointment['service'] ?? '')) : ($appointment['service'] ?? '-');
+    $doctorLabel = function_exists('get_doctor_label') ? get_doctor_label((string)($appointment['doctor'] ?? '')) : ($appointment['doctor'] ?? '-');
+    $dateLabel = function_exists('format_date_label') ? format_date_label((string)($appointment['date'] ?? '')) : ($appointment['date'] ?? '-');
+
     $clinicName = 'Piel en Armonía';
     $subject = 'Cita cancelada - ' . $clinicName;
     $message = "Hola " . ($appointment['name'] ?? 'paciente') . ",\n\n";
     $message .= "Tu cita ha sido cancelada.\n\n";
     $message .= "Detalles de la cita cancelada:\n";
-    $message .= "Servicio: " . get_service_label((string) ($appointment['service'] ?? '-')) . "\n";
-    $message .= "Doctor: " . get_doctor_label((string) ($appointment['doctor'] ?? '-')) . "\n";
-    $message .= "Fecha: " . format_date_label((string) ($appointment['date'] ?? '-')) . "\n";
+    $message .= "Servicio: " . $serviceLabel . "\n";
+    $message .= "Doctor: " . $doctorLabel . "\n";
+    $message .= "Fecha: " . $dateLabel . "\n";
     $message .= "Hora: " . ($appointment['time'] ?? '-') . "\n\n";
-    $message .= "Si deseas reprogramar, visita https://pielarmonia.com/#citas o escribenos por WhatsApp: " . clinic_whatsapp_display() . ".\n\n";
+    $message .= "Si deseas reprogramar, visita https://pielarmonia.com/#citas o escribenos por WhatsApp: +593 98 245 3672.\n\n";
     $message .= "Gracias por confiar en nosotros.";
 
     return send_mail($to, $subject, $message);
@@ -425,13 +241,17 @@ function maybe_send_reminder_email(array $appointment): bool
         return false;
     }
 
+    $serviceLabel = function_exists('get_service_label') ? get_service_label((string)($appointment['service'] ?? '')) : ($appointment['service'] ?? '-');
+    $doctorLabel = function_exists('get_doctor_label') ? get_doctor_label((string)($appointment['doctor'] ?? '')) : ($appointment['doctor'] ?? '-');
+    $dateLabel = function_exists('format_date_label') ? format_date_label((string)($appointment['date'] ?? '')) : ($appointment['date'] ?? '-');
+
     $clinicName = 'Piel en Armonía';
     $subject = 'Recordatorio de cita - ' . $clinicName;
     $body = "Hola " . ($appointment['name'] ?? 'paciente') . ",\n\n";
     $body .= "Te recordamos que tienes una cita programada para mañana.\n\n";
-    $body .= "Servicio: " . get_service_label((string) ($appointment['service'] ?? '-')) . "\n";
-    $body .= "Doctor: " . get_doctor_label((string) ($appointment['doctor'] ?? '-')) . "\n";
-    $body .= "Fecha: " . format_date_label((string) ($appointment['date'] ?? '-')) . "\n";
+    $body .= "Servicio: " . $serviceLabel . "\n";
+    $body .= "Doctor: " . $doctorLabel . "\n";
+    $body .= "Fecha: " . $dateLabel . "\n";
     $body .= "Hora: " . ($appointment['time'] ?? '-') . "\n\n";
 
     $token = $appointment['rescheduleToken'] ?? '';
@@ -442,7 +262,7 @@ function maybe_send_reminder_email(array $appointment): bool
 
     $body .= "Te esperamos. ¡Gracias por confiar en nosotros!\n";
     $body .= "- Equipo Piel en Armonía\n";
-    $body .= "WhatsApp: " . clinic_whatsapp_display();
+    $body .= "WhatsApp: +593 98 245 3672";
 
     return send_mail($to, $subject, $body);
 }
@@ -454,13 +274,17 @@ function maybe_send_reschedule_email(array $appointment): bool
         return false;
     }
 
+    $serviceLabel = function_exists('get_service_label') ? get_service_label((string)($appointment['service'] ?? '')) : ($appointment['service'] ?? '-');
+    $doctorLabel = function_exists('get_doctor_label') ? get_doctor_label((string)($appointment['doctor'] ?? '')) : ($appointment['doctor'] ?? '-');
+    $dateLabel = function_exists('format_date_label') ? format_date_label((string)($appointment['date'] ?? '')) : ($appointment['date'] ?? '-');
+
     $clinicName = 'Piel en Armonía';
     $subject = 'Cita reprogramada - ' . $clinicName;
     $body = "Hola " . ($appointment['name'] ?? 'paciente') . ",\n\n";
     $body .= "Tu cita ha sido reprogramada exitosamente.\n\n";
-    $body .= "Servicio: " . get_service_label((string) ($appointment['service'] ?? '-')) . "\n";
-    $body .= "Doctor: " . get_doctor_label((string) ($appointment['doctor'] ?? '-')) . "\n";
-    $body .= "Nueva fecha: " . format_date_label((string) ($appointment['date'] ?? '-')) . "\n";
+    $body .= "Servicio: " . $serviceLabel . "\n";
+    $body .= "Doctor: " . $doctorLabel . "\n";
+    $body .= "Nueva fecha: " . $dateLabel . "\n";
     $body .= "Nueva hora: " . ($appointment['time'] ?? '-') . "\n\n";
 
     $token = $appointment['rescheduleToken'] ?? '';
