+++ lib/models.php
@@ -1,104 +1,13 @@
 <?php
 declare(strict_types=1);
 
+require_once __DIR__ . '/common.php';
+require_once __DIR__ . '/validation.php';
+require_once __DIR__ . '/business.php';
+
 /**
  * Business logic and data models.
  */
-
-function map_callback_status(string $status): string
-{
-    $normalized = strtolower(trim($status));
-    if ($normalized === 'contacted') {
-        return 'contactado';
-    }
-    if ($normalized === 'pending') {
-        return 'pendiente';
-    }
-    return in_array($normalized, ['pendiente', 'contactado'], true) ? $normalized : 'pendiente';
-}
-
-function map_appointment_status(string $status): string
-{
-    $normalized = strtolower(trim($status));
-    return in_array($normalized, ['confirmed', 'pending', 'cancelled', 'completed'], true)
-        ? $normalized
-        : 'confirmed';
-}
-
-function normalize_review(array $review): array
-{
-    $rating = isset($review['rating']) ? (int) $review['rating'] : 0;
-    if ($rating < 1) {
-        $rating = 1;
-    }
-    if ($rating > 5) {
-        $rating = 5;
-    }
-    return [
-        'id' => isset($review['id']) ? (int) $review['id'] : (int) round(microtime(true) * 1000),
-        'name' => truncate_field(isset($review['name']) ? trim((string) $review['name']) : '', 100),
-        'rating' => $rating,
-        'text' => truncate_field(isset($review['text']) ? trim((string) $review['text']) : '', 2000),
-        'date' => isset($review['date']) ? (string) $review['date'] : local_date('c'),
-        'verified' => isset($review['verified']) ? parse_bool($review['verified']) : true
-    ];
-}
-
-function normalize_callback(array $callback): array
-{
-    return [
-        'id' => isset($callback['id']) ? (int) $callback['id'] : (int) round(microtime(true) * 1000),
-        'telefono' => truncate_field(sanitize_phone((string) ($callback['telefono'] ?? '')), 20),
-        'preferencia' => truncate_field((string) ($callback['preferencia'] ?? ''), 200),
-        'fecha' => isset($callback['fecha']) ? (string) $callback['fecha'] : local_date('c'),
-        'status' => map_callback_status((string) ($callback['status'] ?? 'pendiente'))
-    ];
-}
-
-function get_vat_rate(): float
-{
-    $raw = getenv('PIELARMONIA_VAT_RATE');
-    if (!is_string($raw) || trim($raw) === '') {
-        return 0.12;
-    }
-
-    $rate = (float) trim($raw);
-    if ($rate > 1.0 && $rate <= 100.0) {
-        $rate = $rate / 100.0;
-    }
-
-    if ($rate < 0.0) {
-        return 0.0;
-    }
-    if ($rate > 1.0) {
-        return 1.0;
-    }
-    return $rate;
-}
-
-function get_service_price_amount(string $service): float
-{
-    $prices = [
-        'consulta' => 40.00,
-        'telefono' => 25.00,
-        'video' => 30.00,
-        'laser' => 150.00,
-        'rejuvenecimiento' => 120.00
-    ];
-    return isset($prices[$service]) ? (float) $prices[$service] : 0.0;
-}
-
-function get_service_price(string $service): string
-{
-    return '$' . number_format(get_service_price_amount($service), 2, '.', '');
-}
-
-function get_service_total_price(string $service): string
-{
-    $subtotal = get_service_price_amount($service);
-    $total = $subtotal + ($subtotal * get_vat_rate());
-    return '$' . number_format($total, 2, '.', '');
-}
 
 function get_service_label(string $service): string
 {
@@ -157,6 +66,36 @@ function format_date_label(string $date): string
     $month = (int) date('n', $ts) - 1;
     $year = date('Y', $ts);
     return ucfirst($dias[$dow]) . ' ' . $day . ' de ' . $meses[$month] . ' de ' . $year;
+}
+
+function normalize_review(array $review): array
+{
+    $rating = isset($review['rating']) ? (int) $review['rating'] : 0;
+    if ($rating < 1) {
+        $rating = 1;
+    }
+    if ($rating > 5) {
+        $rating = 5;
+    }
+    return [
+        'id' => isset($review['id']) ? (int) $review['id'] : (int) round(microtime(true) * 1000),
+        'name' => truncate_field(isset($review['name']) ? trim((string) $review['name']) : '', 100),
+        'rating' => $rating,
+        'text' => truncate_field(isset($review['text']) ? trim((string) $review['text']) : '', 2000),
+        'date' => isset($review['date']) ? (string) $review['date'] : local_date('c'),
+        'verified' => isset($review['verified']) ? parse_bool($review['verified']) : true
+    ];
+}
+
+function normalize_callback(array $callback): array
+{
+    return [
+        'id' => isset($callback['id']) ? (int) $callback['id'] : (int) round(microtime(true) * 1000),
+        'telefono' => truncate_field(sanitize_phone((string) ($callback['telefono'] ?? '')), 20),
+        'preferencia' => truncate_field((string) ($callback['preferencia'] ?? ''), 200),
+        'fecha' => isset($callback['fecha']) ? (string) $callback['fecha'] : local_date('c'),
+        'status' => map_callback_status((string) ($callback['status'] ?? 'pendiente'))
+    ];
 }
 
 function normalize_appointment(array $appointment): array
@@ -221,29 +160,4 @@ function normalize_appointment(array $appointment): array
             : bin2hex(random_bytes(16)),
         'reminderSentAt' => truncate_field(trim((string) ($appointment['reminderSentAt'] ?? '')), 30)
     ];
-}
-
-function appointment_slot_taken(array $appointments, string $date, string $time, ?int $excludeId = null, string $doctor = ''): bool
-{
-    foreach ($appointments as $appt) {
-        $id = isset($appt['id']) ? (int) $appt['id'] : null;
-        if ($excludeId !== null && $id === $excludeId) {
-            continue;
-        }
-        $status = map_appointment_status((string) ($appt['status'] ?? 'confirmed'));
-        if ($status === 'cancelled') {
-            continue;
-        }
-        if ((string) ($appt['date'] ?? '') !== $date || (string) ($appt['time'] ?? '') !== $time) {
-            continue;
-        }
-        if ($doctor !== '' && $doctor !== 'indiferente') {
-            $apptDoctor = (string) ($appt['doctor'] ?? '');
-            if ($apptDoctor !== '' && $apptDoctor !== 'indiferente' && $apptDoctor !== $doctor) {
-                continue;
-            }
-        }
-        return true;
-    }
-    return false;
 }
